convert gameobject and dependents to cpp, not using a class yet
